+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Refactored the train function into smaller functions [#249](https://github.com/EarthyScience/EasyHybrid.jl/pull/249)
 - Added a custom features+code sections in index.md [#248](https://github.com/EarthyScience/EasyHybrid.jl/pull/248)
 - Added a custom footer section [#247](https://github.com/EarthyScience/EasyHybrid.jl/pull/247)